Linear interpolation resampling filter working
I've written a basic linear interpolator for the resampler algorithm. it
seems to work pretty well, but there are some odd noise patterns added
to channels in the mid range. I'm not sure if it's due to a bug of some
kind, or the simple linear kernel introducing some images of the signal
or something. I'm also thinkng it could be from a very slow drift in
clock phase since it's running open loop right now.