SFO estimation and Correction working + basic equalizer implimented
The SFO estimation happens to two main stages, the first stage is rough
estimation using the preamble symbols. The first two symbols, including
the first symmetric symbol used for timing detection, are split in half
and DFT'd. The phase difference between the halves is used to estimate
the SFO. Then again for the duplicated symbols 2 and 3, but between the
whole symbols instead of just their halves.

The second stage is done during data transmission using phase
information from pilot channels, which right now are unchanging, but in
the future I'd like to have modulated with a pseudo random sequence.

With this sytem, it can usually get from like 50ppm to within 1 or 0.5ppm using just the
preamble, then within 40 or so symbols it can get within 0.05ppm or
less, and after 50 or 60 symbols it's within much less than 0.01ppm
error.