Found the nagging pilot based SFO estimator bug
It was throwing off the first sample wildly. Issue was, at some point
during reorganization, I must have deleted the line that captures the
value of pilot symbols during the very first data symbol. Then, when the
next data symbol is run, and the first SFO estimation is made based on
the pilot signals, it's always just garbage. Now it works perfectly.

Another issue was during the preamble, the SFO estimations were being
ruined by the changing channel estimation. I moved the channel
estimation so that it only affects the data channels. Hopefully that
doesn't cause me issues later on.

Another problem was the hacky noisy channel rejection statement I wrote
caused the sfo estimate to become nan sometimes. I removed the hacky code
and added a check for divide by zero in case the normalizationFactor
is zero at any point.

There was another issue that was causing wild fluctuation of the SFO
estimation signal. Maybe this was simply due to a large uncorrected SFO
or maybe it was due equalization of the pilot channels. Not really sure