IT TRANSMITTED DATA!!! implemented huffman Decoding
the transmitter pulls data in from a file, XORs it with the random
number sequence I have going and modulates it

The receiver get IQ symbols, quantizes them to constellation indicies,
reverses the huffman coding and XORs it with the same random number to
recover the original data stream that it writes to a file

I'm also getting better at making helper functions to split up larger
tasks.

other bugs:
fixed makefile Dependancy bug that caused updating the header not to
force recompiling of all sources dependant on that header.